Fix percent of code, add code quality info to overview table

+++ core/src/main/kotlin/net/ntworld/codeCleaner/statistic/CodeStatistic.kt
@@ -32,7 +32,7 @@ class CodeStatistic {
                 CodeStatisticLanguageDataImpl(
                     language = it.key.displayName,
                     lines = it.value,
-                    percent = (it.value / total) * 100f
+                    percent = (it.value.toFloat() / total) * 100f
                 )
             }
         )
